# docs/swagger.yaml
basePath: /
definitions:
  github_com_desen_track-me_internal_dto.APIResponse:
    properties:
      data: {}
      errors:
        items:
          type: string
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  github_com_desen_track-me_internal_dto.BlockUserRequest:
    properties:
      user_id:
        type: string
    required:
    - user_id
    type: object
...
  github_com_desen_track-me_internal_dto.FriendRequestResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      receiver:
        $ref: '#/definitions/github_com_desen_track-me_internal_dto.UserResponse'
      sender:
        $ref: '#/definitions/github_com_desen_track-me_internal_dto.UserResponse'
      status:
        type: string
    type: object
...
  github_com_desen_track-me_internal_dto.RespondFriendRequest:
    properties:
      accept:
        type: boolean
      request_id:
        type: string
    required:
    - request_id
    type: object
  github_com_desen_track-me_internal_dto.SendFriendRequest:
    properties:
      receiver_email:
        type: string
    required:
    - receiver_email
    type: object
...
  github_com_desen_track-me_internal_dto.UserResponse:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      phone:
        type: string
      status_message:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Backend REST & WebSocket API for lightweight, privacy-focused mutual
    location sharing.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Mutual Location Sharing API
  version: "1.0"
paths:
...
  /api/v1/friend/block:
    post:
      consumes:
      - application/json
      description: Block a user, remove friendship, and revoke any location sharing
      parameters:
      - description: User ID to block
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_desen_track-me_internal_dto.BlockUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User blocked successfully
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Block a user
      tags:
      - Friend
  /api/v1/friend/blocked:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all users blocked by the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_desen_track-me_internal_dto.UserResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Get blocked users list
      tags:
      - Friend
  /api/v1/friend/list:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all friends of the user
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_desen_track-me_internal_dto.UserResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Get friends list
      tags:
      - Friend
  /api/v1/friend/remove/{id}:
    delete:
      consumes:
      - application/json
      description: Terminate friendship with a user by ID
      parameters:
      - description: Friend User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Friend removed successfully
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Remove a friend
      tags:
      - Friend
  /api/v1/friend/request:
    post:
      consumes:
      - application/json
      description: Initiate a friendship request to another user by email
      parameters:
      - description: Friend request payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_desen_track-me_internal_dto.SendFriendRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Request sent
          schema:
            allOf:
            - $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_desen_track-me_internal_dto.FriendRequestResponse'
              type: object
        "400":
          description: Bad request (self friending or duplicate)
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Send friend request
      tags:
      - Friend
  /api/v1/friend/requests:
    get:
      consumes:
      - application/json
      description: List all incoming pending friend requests for the user
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_desen_track-me_internal_dto.FriendRequestResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Get pending requests
      tags:
      - Friend
  /api/v1/friend/respond:
    post:
      consumes:
      - application/json
      description: Accept or reject a pending friend request
      parameters:
      - description: Response payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_desen_track-me_internal_dto.RespondFriendRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Request responded successfully
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "400":
          description: Validation or logical error
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Respond to friend request
      tags:
      - Friend
  /api/v1/friend/unblock/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a user from your block list
      parameters:
      - description: Blocked User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User unblocked successfully
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Unblock a user
      tags:
      - Friend
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token (e.g. Bearer <token>)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"